move the QueryLogger class to another file
