+++ src/marqo/tensor_search/api.py
@@ -29,6 +29,7 @@ from marqo.core import exceptions as core_exceptions
 from marqo.core.index_management.index_management import IndexManagement
 from marqo.core.inference.api import exceptions as inference_exceptions
 from marqo.core.monitoring import memory_profiler
+from marqo.core.search.query_logger import QueryLogger
 from marqo.inference.inference_cache.caching_inference import CachingInference
 from marqo.inference.native_inference.remote.client.inference_client import NativeInferenceClient
 from marqo.inference.native_inference.remote.client.model_manager_client import ModelManagerClient
@@ -392,35 +393,6 @@ def get_index_stats(index_name: str, marqo_config: config.Config = Depends(get_c
             'storageUsedPercentage': stats.backend.storage_used_percentage
         }
     }
-
-
-# TODO move this inside QueryLogger class and find a way to test
-marqo_query_logger = get_logger('marqo_query')
-
-
-# TODO Move this class to a new file
-class QueryLogger:
-    def __init__(self, query_dict):
-        self.query_dict = query_dict
-        self.error_logged = False
-
-        # TODO find a way to move these two at class level and test it
-        self.slow_query_threshold_ms = float(utils.read_env_vars_and_defaults(EnvVars.MARQO_SLOW_QUERY_THRESHOLD_MS))
-        self.log_query_details = utils.read_env_vars_and_defaults(EnvVars.MARQO_LOG_QUERY_DETAILS).upper() == "TRUE"
-
-    @property
-    def sanitised_query(self):
-        return self.query_dict # TODO sanitise it
-
-    def log_error_query(self, error_message: str):
-        if self.log_query_details:
-            marqo_query_logger.error(f'Failed search query: Error: {error_message}. Query: {self.sanitised_query}')
-            self.error_logged = True  # Mark that error was logged
-
-    def log_slow_query(self, elapsed_time_ms: float):
-        if self.log_query_details and not self.error_logged and elapsed_time_ms >= self.slow_query_threshold_ms:
-            marqo_query_logger.warning(f'Slow search query detected: {elapsed_time_ms:.1f}ms. '
-                                       f'Query: {self.sanitised_query}')
 
 
 @app.post("/indexes/{index_name}/search")
